Started making my switches and got the right mcu symbol
Ummmm.... alot has happened. Finally got the right MCU symbol for the schematic and decided to move ahead and start laying the switches down. Still deciding on what I want. The rotartty encoder with a switch is templting because I can use it to contol sound volume and also as a pause and play button.

Clearly I need to get this show on the read with only 11 days left  to finish it. Lets GOOOOOOOOOOOO